NEWSLHTMLD-2: Refactor newsletter-htmldocs module to newsletter-blog
