+++ src/java/fr/paris/lutece/plugins/newsletter/modules/blog/resources/blog_messages_fr.properties
