+++ src/sql/plugins/newsletter/modules/blog/plugin/create_db_newsletter_blog.sql
