+++ webapp/WEB-INF/conf/plugins/newsletter-blog_context.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE beans PUBLIC "-//SPRING//DTD BEAN//EN" "http://www.springframework.org/dtd/spring-beans.dtd">
+
+<beans>
+	<bean id="newsletter-blog.newsletterHtmlDocDAO" class="fr.paris.lutece.plugins.newsletter.modules.blog.business.NewsletterHtmlDocDAO" />
+	
+	<bean id="newsletter-blog.newsHtmlDocTopicService" class="fr.paris.lutece.plugins.newsletter.modules.blog.service.NewsletterHtmlDocTopicService" />
+</beans>
