+++ webapp/WEB-INF/templates/admin/plugins/newsletter/modify_newsletter_htmldocs_topic_config.html
@@ -1,19 +1,19 @@
 <div class="form-group">
-	<label class="control-label col-xs-12 col-sm-3" for="use_categories">#i18n{module.newsletter.document.modify_document_topic.useCategories} *</label>
+	<label class="control-label col-xs-12 col-sm-3" for="use_categories">#i18n{module.newsletter.blog.modify_document_topic.useCategories} *</label>
 	<div class="col-xs-12 col-sm-9">
 		<label class="radio-inline">
 			<input type="radio"  id="use_categories" name="use_categories" value="true" <#if newsletterDocument.useDocumentCategories>checked="checked"</#if> <#if !category_list?has_content>disabled="disabled"</#if>>
-			#i18n{module.newsletter.document.modify_document_topic.radioUseCategories}
+			#i18n{module.newsletter.blog.modify_document_topic.radioUseCategories}
 		</label>
 		<label class="radio-inline">
 			<input type="radio" id="use_portlets" name="use_categories" value="false" <#if !newsletterDocument.useDocumentCategories>checked="checked"</#if> <#if !document_list_portlets?has_content>disabled="disabled"</#if>>
-			#i18n{module.newsletter.document.modify_document_topic.radioUsePortlets}
+			#i18n{module.newsletter.blog.modify_document_topic.radioUsePortlets}
 		</label>
-		<span class="help-block">#i18n{module.newsletter.document.modify_document_topic.useCategoriesHelp}</span>
+		<span class="help-block">#i18n{module.newsletter.blog.modify_document_topic.useCategoriesHelp}</span>
 	</div>
 </div>
 <div id="categories" class="form-group">
-	<label class="control-label col-xs-12 col-sm-3" for="category_list_id">#i18n{module.newsletter.document.modify_document_topic.categoryList} *</label>
+	<label class="control-label col-xs-12 col-sm-3" for="category_list_id">#i18n{module.newsletter.blog.modify_document_topic.categoryList} *</label>
 	<div class="col-xs-12 col-sm-9">
 		<#if category_list?has_content>
 			<select multiple="multiple" class="form-control" id="category_list_id" name="category_list_id">
@@ -26,12 +26,12 @@
 				</#list>
 			</select>
 		<#else>
-			#i18n{module.newsletter.document.modify_document_topic.labelNoCategory}
+			#i18n{module.newsletter.blog.modify_document_topic.labelNoCategory}
 		</#if>
 	</div>
 </div>
 <div class="form-group" id="documents" >
-	<label class="control-label col-xs-12 col-sm-3" for="portlets_id">#i18n{module.newsletter.document.modify_document_topic.portletList} *</label>
+	<label class="control-label col-xs-12 col-sm-3" for="portlets_id">#i18n{module.newsletter.blog.modify_document_topic.portletList} *</label>
 	<div class="col-xs-12 col-sm-9">
 		<#if document_list_portlets?has_content>
 			<select multiple="multiple" class="form-control" id="portlets_id" name="portlets_id">
@@ -44,12 +44,12 @@
 				</#list>
 			</select>
 		<#else>
-			<div class="callout callout-warning">#i18n{module.newsletter.document.modify_document_topic.labelNoPortlet}</div>
+			<div class="callout callout-warning">#i18n{module.newsletter.blog.modify_document_topic.labelNoPortlet}</div>
 		</#if>
 	</div>
 </div>
 <div class="form-group">
-	<label class="control-label col-xs-12 col-sm-3 for="template_id">#i18n{module.newsletter.document.modify_document_topic.templatesList} *:</label>
+	<label class="control-label col-xs-12 col-sm-3 for="template_id">#i18n{module.newsletter.blog.modify_document_topic.templatesList} *:</label>
 	<div class="col-xs-12 col-sm-9">
 		<#if templates_list?has_content>
 			<#list templates_list as template >
@@ -58,7 +58,7 @@
 					<img src="${img_path}/${template.picture}" class="img-thumbnail img-responsive thumbnail-xs" alt="${template.description}" title="${template.description}"> ${template.description}
 				</label>
 			</#list>
-			<span class="help-text">#i18n{module.newsletter.document.modify_document_topic.templatesHelp}</span>
+			<span class="help-text">#i18n{module.newsletter.blog.modify_document_topic.templatesHelp}</span>
 		<#else>
 			<input type="hidden" name="template_id" value="0">
 			<div class="callout callout-warning">
