+++ webapp/WEB-INF/templates/admin/plugins/newsletter/modules/blog/modify_newsletter_htmldocs_topic_config.html
@@ -1,17 +1,17 @@
 <div class="form-group">
-	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="use_categories">#i18n{module.newsletter.htmldocs.modify_document_topic.useCategories} *</label>
+	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="use_categories">#i18n{module.newsletter.blog.modify_document_topic.useCategories} *</label>
 	<div class="col-xs-12 col-sm-10 col-md-10">
 		<label class="radio-inline">
-			<input type="radio"  id="use_categories" name="use_categories" value="true" <#if newsletterDocument.useDocumentTags>checked="checked"</#if>> #i18n{module.newsletter.htmldocs.modify_document_topic.radioUseTags}
+			<input type="radio"  id="use_categories" name="use_categories" value="true" <#if newsletterDocument.useDocumentTags>checked="checked"</#if>> #i18n{module.newsletter.blog.modify_document_topic.radioUseTags}
 		</label>
 		<label class="radio-inline">
-			<input type="radio" id="use_portlets" name="use_categories" value="false" <#if !newsletterDocument.useDocumentTags>checked="checked"</#if>>#i18n{module.newsletter.htmldocs.modify_document_topic.radioUsePortlets}
+			<input type="radio" id="use_portlets" name="use_categories" value="false" <#if !newsletterDocument.useDocumentTags>checked="checked"</#if>>#i18n{module.newsletter.blog.modify_document_topic.radioUsePortlets}
 		</label>
 		<p class="help-text">#i18n{module.newsletter.htmldocs.modify_document_topic.useCategoriesHelp}</p>
 	</div>
 </div>
 <div id="category" class="form-group">
-	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="category_list_id">#i18n{module.newsletter.htmldocs.modify_document_topic.categoryList} *</label>
+	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="category_list_id">#i18n{module.newsletter.blog.modify_document_topic.categoryList} *</label>
 	<div class="col-xs-12 col-sm-10 col-md-10">
 		<#if category_list?has_content>
 			<select multiple="multiple" class="form-control" id="category_list_id" name="category_list_id" onchange="javascript:clearSelection(document.getElementById('portlets_id'));$('#use_categories').attr('checked', true);" >
@@ -24,12 +24,12 @@
 				</#list>
 			</select>
 		<#else>
-			#i18n{module.newsletter.htmldocs.modify_document_topic.labelNoCategory}
+			#i18n{module.newsletter.blog.modify_document_topic.labelNoCategory}
 		</#if>
 	</div>
 </div>
 <div id="document" class="form-group">
-	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="portlets_id">#i18n{module.newsletter.htmldocs.modify_document_topic.portletList} *</label>
+	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="portlets_id">#i18n{module.newsletter.blog.modify_document_topic.portletList} *</label>
 	<div class="col-xs-12 col-sm-10 col-md-10">
 		<#if document_list_portlets?has_content>
 			<select multiple="multiple" class="form-control" id="portlets_id" name="portlets_id" onchange="javascript:clearSelection(document.getElementById('category_list_id'));$('#use_portlets').attr('checked', true);">
@@ -47,7 +47,7 @@
 	</div>
 </div>
 <div class="form-group">
-	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="template_id">#i18n{module.newsletter.htmldocs.modify_document_topic.templatesList} *:</label>
+	<label class="control-label col-xs-12 col-sm-2 col-md-2" for="template_id">#i18n{module.newsletter.blog.modify_document_topic.templatesList} *:</label>
 	<div class="col-xs-12 col-sm-10 col-md-10">
 		<#if templates_list?has_content>
 			<#list templates_list as template >
@@ -56,7 +56,7 @@
 					<!-- <img src="${img_path}/${template.picture}" class="thumblist-medium" alt="${template.description}" title="${template.description}"> -->
 				</label>
 			</#list>
-			<span class="help-text">#i18n{module.newsletter.htmldocs.modify_document_topic.templatesHelp}</span>
+			<span class="help-text">#i18n{module.newsletter.blog.modify_document_topic.templatesHelp}</span>
 		<#else>
 			<input type="hidden" name="template_id" value="0">
 			<div class="callout callout-warning">
