+++ webapp/WEB-INF/templates/admin/plugins/newsletter/modules/blog/select_htmldocs.html
@@ -1,38 +1,38 @@
 <#include "/admin/insert/insert_utils_macros.html" />
 
-<form class="form-horizontal" method="post" action="jsp/admin/plugins/newsletter/modules/document/DoSearchDocuments.jsp">
+<form class="form-horizontal" method="post" action="jsp/admin/plugins/newsletter/modules/blog/DoSearchDocuments.jsp">
 	<fieldset>
-		<legend>#i18n{module.newsletter.document.select_documents.labelSearchCriteria}</legend>
+		<legend>#i18n{module.newsletter.blog.select_documents.labelSearchCriteria}</legend>
 	    <input type="hidden" name="input" value="${input}" />
 	   	<div class="form-group">
-			<label class="control-label col-xs-12 col-sm-2 col-md-2" for="document_list_id">#i18n{module.newsletter.document.select_documents.labelDocumentListCollection} *</label>
+			<label class="control-label col-xs-12 col-sm-2 col-md-2" for="document_list_id">#i18n{module.newsletter.blog.select_documents.labelDocumentListCollection} *</label>
 			<div class="col-xs-12 col-sm-10 col-md-10">
 				<@comboWithParams name="document_list_id" default_value=document_list_id additionalParameters=" class=\"form-control\"" items=documents_lists_list />
 			</div>
 		</div>
 		<div class="form-group">
-			<label class="control-label col-xs-12 col-sm-2 col-md-2" for="published_date">#i18n{module.newsletter.document.select_documents.labelPublishedSince} *</label>
+			<label class="control-label col-xs-12 col-sm-2 col-md-2" for="published_date">#i18n{module.newsletter.blog.select_documents.labelPublishedSince} *</label>
 	    	<div class="col-xs-12 col-sm-10 col-md-10">
 				<input name="published_date" value="${start_published_date}" class="form-control" type="text">
 			</div>
 		</div>
 		<div class="form-group">
 			<div class="col-xs-12 col-sm-offset-2col-sm-10  col-md-offset-2 col-md-10">
-	    		<button class="btn btn-primary btn-flat" type="submit" >#i18n{module.newsletter.document.select_documents.buttonSearch}</button>
+	    		<button class="btn btn-primary btn-flat" type="submit" >#i18n{module.newsletter.blog.select_documents.buttonSearch}</button>
 	    	</div>
 	    </div>
 	</fieldset>
 </form>
 
-<form class="form-horizontal" method="post" action="jsp/admin/plugins/newsletter/modules/document/DoInsertDocuments.jsp" name="insertDocuments" >
+<form class="form-horizontal" method="post" action="jsp/admin/plugins/newsletter/modules/blog/DoInsertDocuments.jsp" name="insertDocuments" >
 	<fieldset>
-		<legend>#i18n{module.newsletter.document.select_documents.labelDocumentList}</legend>
+		<legend>#i18n{module.newsletter.blog.select_documents.labelDocumentList}</legend>
 	    <input type="hidden" name="input" value="${input}" />
 	    <table class="table table-striped table-condensed">
 			<tr>
 	            <th></th>
-	            <th>#i18n{module.newsletter.document.select_documents.labelDocumentDate}</th>
-	            <th>#i18n{module.newsletter.document.select_documents.labelDocumentAdd}</th>
+	            <th>#i18n{module.newsletter.blog.select_documents.labelDocumentDate}</th>
+	            <th>#i18n{module.newsletter.blog.select_documents.labelDocumentAdd}</th>
 	        </tr>
 	        <#list document_list as document>
 	        <tr>
@@ -46,7 +46,7 @@
 	    </table>
 	    <@comboWithParams name="template_id" default_value=0 additionalParameters=" class=\"form-control\"" items=documents_templates_list />
 		<button class="btn btn-primary btn-xs btn-flat" type="submit">
-			#i18n{module.newsletter.document.select_documents.buttonValidate}
+			#i18n{module.newsletter.blog.select_documents.buttonValidate}
 		</button>
 	</fieldset>
 </form>
