+++ webapp/jsp/admin/plugins/newsletter/modules/blog/DoSearchDocuments.jsp
@@ -1,7 +1,7 @@
 <%@ page errorPage="../../../../ErrorPage.jsp" %>
 <jsp:include page="../../../../insert/InsertServiceHeader.jsp" />
 
-<jsp:useBean id="newsletterService" scope="session" class="fr.paris.lutece.plugins.newsletter.modules.document.web.NewsletterDocumentServiceJspBean" />
+<jsp:useBean id="newsletterService" scope="session" class="fr.paris.lutece.plugins.newsletter.modules.blog.web.NewsletterDocumentServiceJspBean" />
 
 <%= newsletterService.doSearchDocuments( request ) %>
